Add db interactions for actors

+++ cmd/api/movies.go
@@ -9,11 +9,11 @@ import (
 // Добавляет информацию о фильме в базу данных.
 func (app *application) addMovieHandler(w http.ResponseWriter, r *http.Request) {
 	var input struct {
-		Title       string `json:"title"`
-		Description string `json:"description"`
-		ReleaseDate string `json:"release_date"` // RFC3339
-		Rating      int    `json:"rating"`
-		Actors      []int  `json:"actors"`
+		Title       string  `json:"title"`
+		Description string  `json:"description"`
+		ReleaseDate string  `json:"release_date"` // RFC3339
+		Rating      float32 `json:"rating"`
+		Actors      []int   `json:"actors"`
 	}
 
 	err := app.readJSON(w, r, &input)
@@ -56,11 +56,11 @@ func (app *application) updateMovieHandler(w http.ResponseWriter, r *http.Reques
 	}
 
 	var input struct {
-		Title       *string `json:"title"`
-		Description *string `json:"description"`
-		ReleaseDate *string `json:"release_date"` // RFC3339
-		Rating      *int    `json:"rating"`
-		Actors      []int   `json:"actors"`
+		Title       *string  `json:"title"`
+		Description *string  `json:"description"`
+		ReleaseDate *string  `json:"release_date"` // RFC3339
+		Rating      *float32 `json:"rating"`
+		Actors      []int    `json:"actors"`
 	}
 
 	err = app.readJSON(w, r, &input)
